+++ libs/qiniu-upload/src/uploader/mod.rs
@@ -1,7 +1,7 @@
 mod builder;
-mod field;
-mod file_upload;
 mod payload;
+mod upload_field;
+mod upload_file;
 mod upload_json;
 use std::{collections::HashMap, fmt::Debug};
 
@@ -13,8 +13,7 @@ pub use upload_json::JsonPayload;
 
 pub use self::{
     builder::{ManagedUploader, UploaderBuilder},
-    field::ByteUploader,
-    payload::{FilePayload, PayloadContent, PayloadLocal},
+    payload::{ByteUploader, FilePayload, PayloadContent, PayloadLocal},
 };
 use crate::{error, SecretConfig};
 #[derive(Debug)]
@@ -46,6 +45,13 @@ impl Uploader {
     pub async fn upload(
         &self, payload: impl PayloadLocal + PayloadContent,
     ) -> Result<ResponsePayload, error::Error> {
+        info!(
+            content_type = %payload.content_type(),
+            qiniu.uploader.bucket = payload.bucket(),
+            qiniu.uploader.obj = payload.obj_name(),
+            qiniu.uploader.file = payload.file_name(),
+        );
+
         let auto_uploader = self
             .managers
             .get(payload.bucket())
